+++ src/Schema/Directives/InjectDirective.php
@@ -46,13 +46,13 @@ GRAPHQL;
             throw new DefinitionException("The `inject` directive on {$fieldValue->getParentName()} [{$fieldValue->getFieldName()}] must have a `name` argument");
         }
 
-        $fieldValue->wrapResolver(fn (callable $previousResolver) => function ($rootValue, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($contextAttributeName, $argumentName, $previousResolver) {
+        $fieldValue->wrapResolver(fn (callable $resolver) => function ($rootValue, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($contextAttributeName, $argumentName, $resolver) {
             $valueFromContext = data_get($context, $contextAttributeName);
 
             $argumentSet = $resolveInfo->argumentSet;
             $argumentSet->addValue($argumentName, $valueFromContext);
 
-            return $previousResolver(
+            return $resolver(
                 $rootValue,
                 $argumentSet->toArray(),
                 $context,