+++ src/Schema/Directives/WithDirective.php
@@ -47,7 +47,7 @@ GRAPHQL;
     /**
      * @return \Nuwave\Lighthouse\Execution\ModelsLoader\SimpleModelsLoader
      */
-    protected function modelsLoader($parent, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): ModelsLoader
+    protected function modelsLoader(mixed $parent, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): ModelsLoader
     {
         return new SimpleModelsLoader(
             $this->relation(),